move journald formatting to journald-format output plugin
