+++ lib/plugins/input/journald-upload.js
@@ -4,66 +4,11 @@ const throng = require('throng')
 const extractTokenRegEx = /([0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12})/
 const tokenFormatRegEx = /[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}/
 
-// journald fields for lowercase access
 const __CURSOR = '__CURSOR'
-const __REALTIME_TIMESTAMP = '__realtime_timestamp'
-const __SOURCE_REALTIME_TIMESTAMP = '_source_realtime_timestamp'
-const SYSLOG_IDENTIFIER = 'syslog_identifier'
-const _HOSTNAME = '_hostname'
 const _SYSTEMD_UNIT = '_systemd_unit'
-const PRIORITY = 'priority'
-const SYSLOG_FACILITY = 'syslog_facility'
-const CONTAINER_TAG = 'container_tag'
-const CONTAINER_ID_FULL = 'container_id_full'
-const CONTAINER_NAME = 'container_name'
-const Parser = require('../../parser/parser.js')
-const logParser = new Parser()
-// field mapping for Sematext Common Schema
-const processFields = {
-  '_pid': 'pid',
-  '_uid': 'uid',
-  '_gid': 'gid',
-  '_cmdline': 'cmd',
-  '_systemd_cgroup': 'cgroup'
-}
-// mapping for syslog priority and facility values 
-const SEVERITY = [
-  'emerg',
-  'alert',
-  'crit',
-  'err',
-  'warning',
-  'notice',
-  'info',
-  'debug'
-]
-const FACILITY = [
-  'kern',
-  'user',
-  'mail',
-  'daemon',
-  'auth',
-  'syslog',
-  'lpr',
-  'news',
-  'uucp',
-  'cron',
-  'authpriv',
-  'ftp',
-  'ntp',
-  'logaudit',
-  'logalert',
-  'clock',
-  'local0',
-  'local1',
-  'local2',
-  'local3',
-  'local4',
-  'local5',
-  'local6',
-  'local7'
-]
+const SYSLOG_IDENTIFIER = 'syslog_identifier'
 const TokenBlacklist = require('../../util/tokenBlackList.js')
+
 function JournaldUpload (config, eventEmitter) {
   this.config = config
   this.eventEmitter = eventEmitter
@@ -97,7 +42,6 @@ function JournaldUpload (config, eventEmitter) {
   }
   // blacklist invalid tokens
   let self = this
-  
 }
 
 JournaldUpload.prototype.start = function () {
@@ -112,55 +56,8 @@ JournaldUpload.prototype.start = function () {
 }
 
 JournaldUpload.prototype.stop = function (cb) {
-  cb()
-}
-
-// Transforming fields to Sematext Common Schema https://sematext.com/docs/tags/common-schema/
-// see https://sematext.com/docs/agents/sematext-agent/processes/metadata/
-JournaldUpload.prototype.applySematextCommonSchema = function (log) {
-  // use Sematext common schema os.host = hostname
-  let hostname = log[_HOSTNAME]
-  if (hostname) {
-    log.os = {host: hostname}
-    delete log[_HOSTNAME]
-  }
-  let timestamp = log[__REALTIME_TIMESTAMP] || log[__SOURCE_REALTIME_TIMESTAMP]
-  if (timestamp) {
-    var d = new Date(Number(timestamp) / 1000)
-    if (d instanceof Date && !isNaN(d)) {
-      log['@timestamp'] = d
-    }
-  }
-  let prio = log[PRIORITY]
-  let facility = log[log[SYSLOG_FACILITY]]
-  // handling syslog priorit and facility values 
-  if (prio || facility) {
-    log.facility = FACILITY[facility] || String(facility)
-    log.severity = SEVERITY[prio] || String(prio)
-    delete log[PRIORITY]
-    delete log[log[SYSLOG_FACILITY]]
-  }
-  // handling docker journald-drive fields
-  if (log[CONTAINER_ID_FULL]) {
-    log.container = {
-      id: log[CONTAINER_ID_FULL], 
-      name: log[CONTAINER_NAME], 
-      tag: log[CONTAINER_TAG]
-    }
-    delete log[CONTAINER_NAME]
-    delete log[CONTAINER_TAG]
-    delete log[CONTAINER_ID_FULL]
-  }
-  if (log['_pid']) {
-    log.process = {}
-  } else {
-    return
-  }
-  for (field in processFields) {
-    if (log[field]) {
-      log.process[processFields[field]] = log[field]
-      delete log[field]
-    }
+  if (cb) {
+    cb()
   }
 }
 
@@ -185,29 +82,8 @@ JournaldUpload.prototype.emitEvent = function (log, token) {
     // set index, for elasticsearch output plugin
     context.index = token
   }
-  if (this.config.useSematextCommonSchema) {
-    this.applySematextCommonSchema(log)
-  }
-  let source = log[_SYSTEMD_UNIT]
-  if (log.container) {
-    source = log.container.tag || log.container.name
-  }
-  // optional parsing of message field with logagent patterns
-  if (config.parseMessageField === true) {
-    logParser.parseLine(log.message, source, function (err, data) {
-      if (data && data._type) {
-        let type = `${data._type}`
-        delete data['@timestamp']
-        delete data['logSource']
-        delete data._type
-        log[type] = {}
-        Object.assign(log[type], data)
-      }
-      this.eventEmitter.emit('data.object', log, context)
-    }.bind(this))
-  } else {
-    this.eventEmitter.emit('data.object', log, context)
-  }
+
+  this.eventEmitter.emit('data.object', log, context)
 }
 
 JournaldUpload.prototype.addTags = function (log) {
